Fixed a little startp bug.

+++ 2.4/docker-entrypoint.sh
@@ -86,8 +86,10 @@ if [ "${SSL_CERT:-none}" = "selfsigned" ]; then
     fi
 else
     if [ "x$CERT_PATH" != "x" ] && [ -e "${CERT_PATH}/privkey.pem" ] && [ -e "${CERT_PATH}/cert.pem" ] ; then
-        ln -s "${CERT_PATH}/privkey.pem" "/";
-        ln -s "${CERT_PATH}/cert.pem" "/";
+        if [! -e "/privkey.pem" ] && [! -e "/cert.pem" ] ; then
+            ln -s "${CERT_PATH}/privkey.pem" "/";
+            ln -s "${CERT_PATH}/cert.pem" "/";
+        fi
     fi
 fi
